requirements.txt
fastapi
uvicorn
python-telegram-bot
pydantic
sqlalchemy
databases
aiosqlite
requests
python-multipart
yookassa
jinja2
python-dotenv